refactor JobEvent
